move lazy fetch in cache
