+++ core/services/ocr2/plugins/ccip/internal/cache/lazy.go
@@ -1,9 +1,11 @@
-package ccip
+package cache
 
 import "sync"
 
 type LazyFunction[T any] func() (T, error)
 
+// LazyFetch caches the results during the first call and then returns the cached value
+// on each consecutive call.
 func LazyFetch[T any](fun LazyFunction[T]) LazyFunction[T] {
 	var result T
 	var err error
