+++ core/services/ocr2/plugins/ccip/internal/cache/lazy_test.go
@@ -1,4 +1,4 @@
-package ccip
+package cache
 
 import (
 	"fmt"
